New package browser.analyses
Moved browser.analyses.py and browser.late_analyses.py inside
