+++ bika/lims/browser/analyses/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of SENAITE.CORE
+#
+# Copyright 2018 by it's authors.
+# Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
+
+from .view import AnalysesView
+from .qc import QCAnalysesView
+from .late import LateAnalysesView
