+++ bika/lims/browser/configure.zcml
@@ -24,7 +24,6 @@
   <include file="instrument.zcml"/>
   <include file="instrumenttype.zcml"/>
   <include file="instrumentlocation.zcml"/>
-  <include file="late_analyses.zcml"/>
   <include file="manufacturer.zcml"/>
   <include file="menu.zcml"/>
   <include file="method.zcml"/>
@@ -40,6 +39,7 @@
   <include file="worksheettemplate.zcml"/>
 
   <include package=".aggregatedanalyses"/>
+  <include package=".analyses"/>
   <include package=".analysisrequest"/>
   <include package=".batch"/>
   <include package=".client"/>
